breaking up classes related to different kinds of turret deployables
