+++ src/test/scala/objects/DeployableTest.scala
@@ -597,7 +597,7 @@ class TurretControlConstructTest extends ActorTest {
   "TurretControl" should {
     "construct" in {
       val obj = new TurretDeployable(GlobalDefinitions.spitfire_turret)
-      system.actorOf(Props(classOf[TurretControl], obj), s"${obj.Definition.Name}_test")
+      system.actorOf(Props(classOf[TurretDeployableControl], obj), s"${obj.Definition.Name}_test")
     }
   }
 }
@@ -642,7 +642,7 @@ class TurretControlMountTest extends ActorTest {
         override def SetupNumberPools() = {}
         this.actor = new TestProbe(system).ref.toTyped[ZoneActor.Command]
       }
-      obj.Actor = system.actorOf(Props(classOf[TurretControl], obj), s"${obj.Definition.Name}_test")
+      obj.Actor = system.actorOf(Props(classOf[TurretDeployableControl], obj), s"${obj.Definition.Name}_test")
 
       assert(obj.Seats(0).occupant.isEmpty)
       val player1 = Player(Avatar(0, "test1", PlanetSideEmpire.TR, CharacterSex.Male, 0, CharacterVoice.Mute))
@@ -662,7 +662,7 @@ class TurretControlBlockMountTest extends ActorTest {
     "block mounting by others if already mounted by someone" in {
       val obj = new TurretDeployable(GlobalDefinitions.portable_manned_turret_tr) { GUID = PlanetSideGUID(1) }
       obj.Faction = PlanetSideEmpire.TR
-      obj.Actor = system.actorOf(Props(classOf[TurretControl], obj), s"${obj.Definition.Name}_test")
+      obj.Actor = system.actorOf(Props(classOf[TurretDeployableControl], obj), s"${obj.Definition.Name}_test")
       obj.Zone = new Zone("test", new ZoneMap("test"), 0) {
         override def SetupNumberPools() = {}
         this.actor = new TestProbe(system).ref.toTyped[ZoneActor.Command]
@@ -694,7 +694,7 @@ class TurretControlBlockBetrayalMountTest extends ActorTest {
     "block mounting by players of another faction" in {
       val obj = new TurretDeployable(GlobalDefinitions.portable_manned_turret_tr) { GUID = PlanetSideGUID(1) }
       obj.Faction = PlanetSideEmpire.TR
-      obj.Actor = system.actorOf(Props(classOf[TurretControl], obj), s"${obj.Definition.Name}_test")
+      obj.Actor = system.actorOf(Props(classOf[TurretDeployableControl], obj), s"${obj.Definition.Name}_test")
 
       assert(obj.Seats(0).occupant.isEmpty)
       val player = Player(Avatar(0, "test", PlanetSideEmpire.VS, CharacterSex.Male, 0, CharacterVoice.Mute))
@@ -718,7 +718,7 @@ class TurretControlDismountTest extends ActorTest {
         override def SetupNumberPools() = {}
         this.actor = new TestProbe(system).ref.toTyped[ZoneActor.Command]
       }
-      obj.Actor = system.actorOf(Props(classOf[TurretControl], obj), s"${obj.Definition.Name}_test")
+      obj.Actor = system.actorOf(Props(classOf[TurretDeployableControl], obj), s"${obj.Definition.Name}_test")
 
       assert(obj.Seats(0).occupant.isEmpty)
       val player = Player(Avatar(0, "test", PlanetSideEmpire.TR, CharacterSex.Male, 0, CharacterVoice.Mute))
@@ -758,7 +758,7 @@ class TurretControlBetrayalMountTest extends ActorTest {
         }
       }
       obj.Faction = PlanetSideEmpire.TR
-      obj.Actor = system.actorOf(Props(classOf[TurretControl], obj), s"${obj.Definition.Name}_test")
+      obj.Actor = system.actorOf(Props(classOf[TurretDeployableControl], obj), s"${obj.Definition.Name}_test")
       val probe = new TestProbe(system)
 
       assert(obj.Seats(0).occupant.isEmpty)
